requirements.txt
pandas==2.2.2
numpy==2.1.3
scikit-learn==1.5.2
requests==2.32.3
streamlit==1.39.0
fpdf2==2.7.9
matplotlib==3.9.2
python-dotenv==1.0.1
joblib==1.4.2
